+++ include/tokenizers/ngram/ngram_pos_tokenizer.h
@@ -32,7 +32,7 @@ class ngram_pos_tokenizer: public ngram_simple_tokenizer
         virtual void tokenize_document(
                 index::document & document,
                 std::function<term_id(const std::string &)> mapping,
-                const std::shared_ptr<std::unordered_map<term_id, unsigned int>> & docFreqs = nullptr
+                const std::shared_ptr<std::unordered_map<term_id, uint64_t>> & docFreqs = nullptr
         );
 };
