+++ include/tokenizers/tree/subtree_tokenizer.h
@@ -20,7 +20,7 @@ namespace tokenizers {
 class subtree_tokenizer: public tree_tokenizer<subtree_tokenizer> {
     public:
         using mapping_fn = std::function<term_id(const std::string &)>;
-        using doc_freq_ptr = std::shared_ptr<std::unordered_map<term_id, unsigned int>>;
+        using doc_freq_ptr = std::shared_ptr<std::unordered_map<term_id, uint64_t>>;
 
         /**
          * Counts occurrences of subtrees in this document's parse_trees.
