+++ src/tokenizers/tokenizer.cpp
@@ -26,7 +26,7 @@ tokenizer::tokenizer():
 { /* nothing */ }
 
 void tokenizer::tokenize(index::document & document,
-        const std::shared_ptr<std::unordered_map<term_id, unsigned int>> & doc_freq) {
+        const std::shared_ptr<std::unordered_map<term_id, uint64_t>> & doc_freq) {
     tokenize_document(document, std::bind(&tokenizer::mapping, this, _1), doc_freq);
 }
